start work on Gremlin syntax test suite. Updated Division to always work in doubles.

+++ src/test/java/com/tinkerpop/gremlin/compiler/operations/math/DivisionTest.java
@@ -17,20 +17,20 @@ public class DivisionTest extends BaseTest {
         this.stopWatch();
         operation = new Division(createUnary(4.0f), createUnary(2.0f));
         printPerformance("div operation", 2, "floats", this.stopWatch());
-        assertEquals(operation.compute().getValue().getClass(), Float.class);
-        assertEquals(operation.compute().getValue(), 2.0f);
+        assertEquals(operation.compute().getValue().getClass(), Double.class);
+        assertEquals(operation.compute().getValue(), 2.0d);
 
         this.stopWatch();
         operation = new Division(createUnary(4l), createUnary(2l));
         printPerformance("div operation", 2, "longs", this.stopWatch());
-        assertEquals(operation.compute().getValue().getClass(), Long.class);
-        assertEquals(operation.compute().getValue(), 2l);
+        assertEquals(operation.compute().getValue().getClass(), Double.class);
+        assertEquals(operation.compute().getValue(), 2.0d);
 
         this.stopWatch();
         operation = new Division(createUnary(4), createUnary(2));
         printPerformance("div operation", 2, "integers", this.stopWatch());
-        assertEquals(operation.compute().getValue().getClass(), Integer.class);
-        assertEquals(operation.compute().getValue(), 2);
+        assertEquals(operation.compute().getValue().getClass(), Double.class);
+        assertEquals(operation.compute().getValue(), 2.0d);
     }
 
     public void testDivisionDifferentTypes() {
@@ -44,14 +44,14 @@ public class DivisionTest extends BaseTest {
         this.stopWatch();
         op = new Division(createUnary(4.0f), createUnary(2l));
         printPerformance("div operation", 2, "float/long", this.stopWatch());
-        assertEquals(op.compute().getValue().getClass(), Float.class);
-        assertEquals(op.compute().getValue(), 2.0f);
+        assertEquals(op.compute().getValue().getClass(), Double.class);
+        assertEquals(op.compute().getValue(), 2.0d);
 
         this.stopWatch();
         op = new Division(createUnary(4l), createUnary(2));
         printPerformance("div operation", 2, "long/int", this.stopWatch());
-        assertEquals(op.compute().getValue().getClass(), Long.class);
-        assertEquals(op.compute().getValue(), 2l);
+        assertEquals(op.compute().getValue().getClass(), Double.class);
+        assertEquals(op.compute().getValue(), 2.0d);
 
         this.stopWatch();
         op = new Division(createUnary(4), createUnary(2.0d));
